Back/requirements.txt
django>=3.2.5,<3.3.0
djangorestframework>=3.12.0,<3.13.0
psycopg2>=2.9.0,<2.9.5
Pillow>=8.3.0,<8.4.0
flake8>=3.9.0,<3.9.9
django-phone-field>=1.8.0,<1.8.5
django-import-export>=2.4.8,<2.5.5
django-filter>=2.4.0